microservices/web/requirements.txt
flask
requests
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-sdk
opentelemetry-api